Resolver: avoid signed overflow assembling values from DNS responses
A-record address and TTL bytes are u_char, promoted to (signed) int
in the shift expression; for any first octet >= 128 the << 24
overflows int, which is undefined behaviour on bytes that arrive in a
DNS response (well-formed TTLs keep the top bit clear per RFC 2181,
but a hostile response need not).

Beyond UBSan's shift check making every such resolution a fatal trap
in -fno-sanitize-recover builds, the overflow undermines a
load-bearing check: the "ttl < 0" guard after each TTL assembly is
reachable only through the overflow, so a compiler doing value-range
propagation may conclude ttl >= 0 and delete it.  Without the guard,
a TTL of 0x80000000 would pass through
rn->ttl = ngx_min(rn->ttl, (uint32_t) ttl) and on into
rn->valid = ngx_time() + (time_t) rn->ttl, keeping the entry for
roughly 68 years.

Widen the first term explicitly at all four assembly sites, following
the existing idiom of casting only the leading << 24 term
(ngx_inet.c, ngx_stream_geo_module.c, ngx_stream_access_module.c).
On current gcc and clang the wrapped value produces the intended bit
pattern and the guard fires, so observable behaviour is unchanged.

Closes: https://github.com/nginx/nginx/issues/1676

+++ src/core/ngx_resolver.c
@@ -2194,7 +2194,7 @@ ngx_resolver_process_a(ngx_resolver_t *r, u_char *buf, size_t n,
         type = (an->type_hi << 8) + an->type_lo;
         class = (an->class_hi << 8) + an->class_lo;
         len = (an->len_hi << 8) + an->len_lo;
-        ttl = (an->ttl[0] << 24) + (an->ttl[1] << 16)
+        ttl = ((uint32_t) an->ttl[0] << 24) + (an->ttl[1] << 16)
             + (an->ttl[2] << 8) + (an->ttl[3]);
 
         if (class != 1) {
@@ -2356,7 +2356,7 @@ ngx_resolver_process_a(ngx_resolver_t *r, u_char *buf, size_t n,
 
             if (type == NGX_RESOLVE_A) {
 
-                addr[j] = htonl((buf[i] << 24) + (buf[i + 1] << 16)
+                addr[j] = htonl(((uint32_t) buf[i] << 24) + (buf[i + 1] << 16)
                                 + (buf[i + 2] << 8) + (buf[i + 3]));
 
                 if (++j == naddrs) {
@@ -2736,7 +2736,7 @@ ngx_resolver_process_srv(ngx_resolver_t *r, u_char *buf, size_t n,
         type = (an->type_hi << 8) + an->type_lo;
         class = (an->class_hi << 8) + an->class_lo;
         len = (an->len_hi << 8) + an->len_lo;
-        ttl = (an->ttl[0] << 24) + (an->ttl[1] << 16)
+        ttl = ((uint32_t) an->ttl[0] << 24) + (an->ttl[1] << 16)
             + (an->ttl[2] << 8) + (an->ttl[3]);
 
         if (class != 1) {
@@ -3301,7 +3301,7 @@ valid:
         type = (an->type_hi << 8) + an->type_lo;
         class = (an->class_hi << 8) + an->class_lo;
         len = (an->len_hi << 8) + an->len_lo;
-        ttl = (an->ttl[0] << 24) + (an->ttl[1] << 16)
+        ttl = ((uint32_t) an->ttl[0] << 24) + (an->ttl[1] << 16)
             + (an->ttl[2] << 8) + (an->ttl[3]);
 
         if (class != 1) {
